issues

+++ app/Http/Controllers/ParticipantController.php
@@ -4,23 +4,45 @@ namespace App\Http\Controllers;
 
 use App\Academy;
 use App\City;
+use App\Http\Requests\AdminRequest;
 use App\Role;
 use App\User;
+use Exception;
+use Illuminate\Database\QueryException;
 use Illuminate\Http\Request;
-use App\Http\Requests;
-
+use Illuminate\Support\Facades\Auth;
+use PDOException;
 
 class ParticipantController extends Controller
 {
     /**
      * Display a listing of the resource.
      *
+     * @param AdminRequest $request
      * @return \Illuminate\Http\Response
      */
     public function index()
     {
-        $participants = Role::where('name', 'participant')->first()->users()->get();
-        return view ('admin.participants.index', compact('participants'));
+        try
+        {
+            $participants = Role::where('name', 'participant')->first()->users()->get();
+            return view ('admin.participants.index', compact('participants'));
+        }
+        catch(QueryException $e)
+        {
+            flash('No pudo completarse la operación', 'danger');
+            return redirect('admin/recetas');
+        }
+        catch(PDOException $e)
+        {
+            flash('No pudo completarse la operación', 'danger');
+            return redirect('admin/recetas');
+        }
+        catch(Exception $e)
+        {
+            flash('No pudo completarse la operación', 'danger');
+            return redirect('admin/recetas');
+        }
     }
 
     /**
@@ -52,8 +74,26 @@ class ParticipantController extends Controller
      */
     public function show($id)
     {
-        $participant = User::findOrFail($id);
-        return view ('admin.participants.show', compact('participant'));
+        try
+        {
+            $participant = User::findOrFail($id);
+            return view ('admin.participants.show', compact('participant'));
+        }
+        catch(QueryException $e)
+        {
+            flash('No pudo completarse la operación', 'danger');
+            return redirect('admin/participantes');
+        }
+        catch(PDOException $e)
+        {
+            flash('No pudo completarse la operación', 'danger');
+            return redirect('admin/participantes');
+        }
+        catch(Exception $e)
+        {
+            flash('No pudo completarse la operación', 'danger');
+            return redirect('admin/participantes');
+        }
     }
 
     /**
@@ -64,38 +104,54 @@ class ParticipantController extends Controller
      */
     public function edit($id)
     {
-        $participant = User::findOrFail($id);
+        if (Auth::user()->hasRole('super_admin')) {
+            try {
+                $participant = User::findOrFail($id);
 
-        $cities = City::pluck('name', 'id');
-        $cities->prepend('Seleccione una ciudad', '');
+                $cities = City::pluck('name', 'id');
+                $cities->prepend('Seleccione una ciudad', '');
 
-        $academies = Academy::pluck('name', 'id');
-        $academies->prepend('N/A', 'N/A');
-        $academies->prepend('Seleccione una academia', '');
+                $academies = Academy::pluck('name', 'id');
+                $academies->prepend('N/A', 'N/A');
+                $academies->prepend('Seleccione una academia', '');
 
-        $sizes = array
-        (''=> 'Seleccione una talla',
-            'SS' => 'SS',
-            'S' => 'S',
-            'M' => 'M',
-            'L' => 'L',
-            'XL' => 'XL',
-            'XXL' => 'XXL',
-            'Otro' => 'Otro');
+                $sizes = array
+                ('' => 'Seleccione una talla',
+                    'SS' => 'SS',
+                    'S' => 'S',
+                    'M' => 'M',
+                    'L' => 'L',
+                    'XL' => 'XL',
+                    'XXL' => 'XXL',
+                    'Otro' => 'Otro');
 
-        $categories = array
-        (''=> 'Seleccione una categoría',
-            'Aficionado/Público General' => 'Aficionado/Público General',
-            'Estudiante/Profesional' => 'Estudiante/Profesional');
+                $categories = array
+                ('' => 'Seleccione una categoría',
+                    'Aficionado/Público General' => 'Aficionado/Público General',
+                    'Estudiante/Profesional' => 'Estudiante/Profesional');
 
-        $types = array
-        (''=> 'Seleccione un tipo',
-            'N/A'=>'N/A',
-            'Oficiante' => 'Oficiante',
-            'Estudiante en curso' => 'Estudiante en curso',
-            'Egresado' => 'Egresado');
+                $types = array
+                ('' => 'Seleccione un tipo',
+                    'N/A' => 'N/A',
+                    'Oficiante' => 'Oficiante',
+                    'Estudiante en curso' => 'Estudiante en curso',
+                    'Egresado' => 'Egresado');
 
-        return view ('admin.participants.edit', compact('participant', 'cities', 'academies', 'sizes', 'categories', 'types'));
+                return view('admin.participants.edit', compact('participant', 'cities', 'academies',
+                    'sizes', 'categories', 'types'));
+            } catch (QueryException $e) {
+                flash('No pudo completarse la operación', 'danger');
+                return redirect('admin/participantes');
+            } catch (PDOException $e) {
+                flash('No pudo completarse la operación', 'danger');
+                return redirect('admin/participantes');
+            } catch (Exception $e) {
+                flash('No pudo completarse la operación', 'danger');
+                return redirect('admin/participantes');
+            }
+        }
+        flash('No tiene permiso para realizar la operación', 'danger');
+        return redirect('admin/participantes');
     }
 
     /**
@@ -107,9 +163,32 @@ class ParticipantController extends Controller
      */
     public function update(Request $request, $id)
     {
-        $user = User::findOrFail($id);
-        $user->update($request->all());
-        flash('Participante actualizado exitosamente', 'success');
+        if (Auth::user()->hasRole('super_admin'))
+        {
+            try
+            {
+                $user = User::findOrFail($id);
+                $user->update($request->all());
+                flash('Participante actualizado exitosamente', 'success');
+                return redirect('admin/participantes');
+            }
+            catch (QueryException $e)
+            {
+                flash('No pudo completarse la operación', 'danger');
+                return redirect('admin/participantes');
+            }
+            catch (PDOException $e)
+            {
+                flash('No pudo completarse la operación', 'danger');
+                return redirect('admin/participantes');
+            }
+            catch (Exception $e)
+            {
+                flash('No pudo completarse la operación', 'danger');
+                return redirect('admin/participantes');
+            }
+        }
+        flash('No tiene permiso para realizar la operación', 'danger');
         return redirect('admin/participantes');
     }
 
@@ -121,8 +200,31 @@ class ParticipantController extends Controller
      */
     public function destroy($id)
     {
-        User::destroy($id);
-        flash('Participante eliminado exitosamente', 'success');
+        if (Auth::user()->hasRole('super_admin'))
+        {
+            try
+            {
+                User::destroy($id);
+                flash('Participante eliminado exitosamente', 'success');
+                return redirect('admin/participantes');
+            }
+            catch (QueryException $e)
+            {
+                flash('No pudo completarse la operación', 'danger');
+                return redirect('admin/participantes');
+            }
+            catch (PDOException $e)
+            {
+                flash('No pudo completarse la operación', 'danger');
+                return redirect('admin/participantes');
+            }
+            catch (Exception $e)
+            {
+                flash('No pudo completarse la operación', 'danger');
+                return redirect('admin/participantes');
+            }
+        }
+        flash('No tiene permiso para realizar la operación', 'danger');
         return redirect('admin/participantes');
     }
 }
